Support source information on 1.24 and 1.25
As of 1.25, slog provides a slog.Record.Source method. I want to use
that, but I also want to support the last two versions of Go.

For 1.24, I build the source information manually, while for 1.25 I use
the new method. I wrap the two choices in a local `source` function, and
build tags will handle the proper build.

Note to self: when 1.26 is released, I'll drop the wrapper function and
call Source directly.